moved function to main file
